// confmod-app/frontend/src/app/openapi/model/rawScopeConfigInput.ts
/**
 * Confmod Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RawObservationConfigInput } from './rawObservationConfigInput';


export interface RawScopeConfigInput { 
    outFile?: string;
    observations: { [key: string]: RawObservationConfigInput; };
    metadata?: Array<string>;
    includePayload?: boolean;
}